+++ packages/frontend/src/components/VisualizationConfigs/PieChartConfig/Display/index.tsx
